--wip-- [skip ci]

+++ src/main/scala/com/github/celadari/jsonlogicscala/evaluate/defaults/OperatorMinus.scala
@@ -4,15 +4,15 @@ import com.github.celadari.jsonlogicscala.evaluate.Operator
 
 object OperatorMinus extends Operator {
 
-  def $minus(num1: java.lang.Byte, num2: java.lang.Byte): java.lang.Byte = (num1 - num2).toByte
-  def $minus(num1: java.lang.Byte, num2: java.lang.Short): java.lang.Short = (num1 - num2).toShort
+  def $minus(num1: java.lang.Byte, num2: java.lang.Byte): java.lang.Integer = num1 - num2
+  def $minus(num1: java.lang.Byte, num2: java.lang.Short): java.lang.Integer = num1 - num2
   def $minus(num1: java.lang.Byte, num2: java.lang.Integer): java.lang.Integer = num1 - num2
   def $minus(num1: java.lang.Byte, num2: java.lang.Long): java.lang.Long = num1 - num2
   def $minus(num1: java.lang.Byte, num2: java.lang.Float): java.lang.Float = num1 - num2
   def $minus(num1: java.lang.Byte, num2: java.lang.Double): java.lang.Double = num1 - num2
 
-  def $minus(num1: java.lang.Short, num2: java.lang.Byte): java.lang.Short = (num1 - num2).toShort
-  def $minus(num1: java.lang.Short, num2: java.lang.Short): java.lang.Short = (num1 - num2).toShort
+  def $minus(num1: java.lang.Short, num2: java.lang.Byte): java.lang.Integer = num1 - num2
+  def $minus(num1: java.lang.Short, num2: java.lang.Short): java.lang.Integer = num1 - num2
   def $minus(num1: java.lang.Short, num2: java.lang.Integer): java.lang.Integer = num1 - num2
   def $minus(num1: java.lang.Short, num2: java.lang.Long): java.lang.Long = num1 - num2
   def $minus(num1: java.lang.Short, num2: java.lang.Float): java.lang.Float = num1 - num2
