+++ src/main/scala/com/github/celadari/jsonlogicscala/evaluate/defaults/OperatorTimes.scala
@@ -4,15 +4,15 @@ import com.github.celadari.jsonlogicscala.evaluate.Operator
 
 object OperatorTimes extends Operator {
 
-  def $times(num1: java.lang.Byte, num2: java.lang.Byte): java.lang.Byte = (num1 * num2).toByte
-  def $times(num1: java.lang.Byte, num2: java.lang.Short): java.lang.Short = (num1 * num2).toShort
+  def $times(num1: java.lang.Byte, num2: java.lang.Byte): java.lang.Integer = num1 * num2
+  def $times(num1: java.lang.Byte, num2: java.lang.Short): java.lang.Integer = num1 * num2
   def $times(num1: java.lang.Byte, num2: java.lang.Integer): java.lang.Integer = num1 * num2
   def $times(num1: java.lang.Byte, num2: java.lang.Long): java.lang.Long = num1 * num2
   def $times(num1: java.lang.Byte, num2: java.lang.Float): java.lang.Float = num1 * num2
   def $times(num1: java.lang.Byte, num2: java.lang.Double): java.lang.Double = num1 * num2
 
-  def $times(num1: java.lang.Short, num2: java.lang.Byte): java.lang.Short = (num1 * num2).toShort
-  def $times(num1: java.lang.Short, num2: java.lang.Short): java.lang.Short = (num1 * num2).toShort
+  def $times(num1: java.lang.Short, num2: java.lang.Byte): java.lang.Integer = num1 * num2
+  def $times(num1: java.lang.Short, num2: java.lang.Short): java.lang.Integer = num1 * num2
   def $times(num1: java.lang.Short, num2: java.lang.Integer): java.lang.Integer = num1 * num2
   def $times(num1: java.lang.Short, num2: java.lang.Long): java.lang.Long = num1 * num2
   def $times(num1: java.lang.Short, num2: java.lang.Float): java.lang.Float = num1 * num2
